backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.0",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-ulid>=3.1.0",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "typing-extensions>=4.15.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]
